+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ### Changelog
 
 - taken out concrete file check out of resources loop
+- Handling not finding resource version